+++ src/Common/Database/DbBase.cs
@@ -1,4 +1,5 @@
-﻿using JsonFlatFileDataStore;
+﻿using Common.Observe;
+using JsonFlatFileDataStore;
 using Serilog;
 using System;
 using System.IO;
