+++ src/Garmin/ApiClient.cs
@@ -1,4 +1,5 @@
 ﻿using Common;
+using Common.Observe;
 using Flurl.Http;
 using Garmin.Dto;
 using Serilog;
