+++ src/Peloton/ApiClient.cs
@@ -1,5 +1,6 @@
 ﻿using Common;
 using Common.Dto;
+using Common.Observe;
 using Flurl.Http;
 using Newtonsoft.Json.Linq;
 using Peloton.Dto;
