+++ src/PelotonToGarminConsole/PelotonToGarminConsole.csproj
@@ -6,7 +6,6 @@
     <AssemblyName>PelotonToGarminConsole</AssemblyName>
     <RootNamespace>PelotonToGarminConsole</RootNamespace>
     <RuntimeIdentifiers>win10-x64;osx-x64;linux-x64</RuntimeIdentifiers>
-    <StartupObject>PelotonToGarminConsole.Program</StartupObject>
     <PackageProjectUrl>https://github.com/philosowaffle/peloton-to-garmin</PackageProjectUrl>
     <RepositoryUrl>https://github.com/philosowaffle/peloton-to-garmin</RepositoryUrl>
     <Authors>Bailey Belvis</Authors>
@@ -33,6 +32,7 @@
     <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.10.13" />
     <PackageReference Include="prometheus-net.DotNetRuntime" Version="4.1.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
+	<PackageReference Include="Serilog.AspNetCore" Version="4.1.0" />
     <PackageReference Include="Serilog.Enrichers.Span" Version="1.3.0" />
     <PackageReference Include="Serilog.Settings.Configuration" Version="3.2.0" />
     <PackageReference Include="Serilog.Sinks.Console" Version="4.0.0" />
@@ -43,7 +43,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Sync\Sync.csproj" />
+    <ProjectReference Include="..\Sync\Sync.csproj" />
     <ProjectReference Include="..\Common\Common.csproj" />
     <ProjectReference Include="..\Conversion\Conversion.csproj" />
     <ProjectReference Include="..\Garmin\Garmin.csproj" />
