+++ src/Sync/Sync.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Serilog" Version="2.10.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Common.csproj" />
+    <ProjectReference Include="..\Conversion\Conversion.csproj" />
+    <ProjectReference Include="..\Garmin\Garmin.csproj" />
+    <ProjectReference Include="..\Peloton\Peloton.csproj" />
+  </ItemGroup>
+
+</Project>
